+++ src/app/welcome/welcome.component.ts
@@ -8,6 +8,7 @@ import { TodoFileService } from '../shared/todo-file.service';
 import { FilePathValidator } from '../shared/validators';
 import { openFilePicker } from '../shared/helpers/file-picker';
 import { hideActionBar } from '../shared/helpers/page';
+import { showToast } from '../shared/helpers/toast';
 
 @Component({
     selector: 'ms-welcome',
@@ -44,9 +45,12 @@ export class WelcomeComponent implements OnInit {
 
     openPicker() {
         openFilePicker().then((filePath) => {
-            this.form.controls.filePath.setValue(filePath);
+            if (filePath) {
+                this.form.controls.filePath.setValue(filePath);
+            }
         }).catch((error) => {
             console.warn(error);
+            showToast(error.toString());
         });
     }